fix: db timeout, session expiry, avatar_url column, db-pool null export, dead config

+++ marketplace/src/pages/api/settings/notifications.ts
@@ -11,7 +11,7 @@ async function ensureEmailPrefsSchema() {
   await query(`
     CREATE TABLE IF NOT EXISTS user_email_preferences (
       id SERIAL PRIMARY KEY,
-      user_id INTEGER NOT NULL UNIQUE REFERENCES users(id) ON DELETE CASCADE,
+      user_id BIGINT NOT NULL UNIQUE REFERENCES users(id) ON DELETE CASCADE,
       marketing BOOLEAN DEFAULT FALSE,
       notifications BOOLEAN DEFAULT TRUE,
       product_updates BOOLEAN DEFAULT TRUE,
